Add auth handler to effect
Currently the onAuthStateChanged listener is registered every time the state context is re-rendered, which happens every time the auth state changes, causing an infinite loop. Putting it in a useEffect (that only runs once) prevents this, preventing a bunch of auth state change events and thus a bunch of re-renders.